requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2
django-allauth==0.54.0
django-cors-headers==4.2.0
django-extensions==3.2.1
django-graphql-jwt==0.3.4
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
MarkupSafe==2.1.2
oauthlib==3.2.2
promise==2.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyotp==2.9.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.2
whitenoise==6.4.0
